Fix issue with unicode chars
There was an issue with rg/ag and unicode chars, as rg/ag would ignore
unicode and count one byte as one column. When moving using cursor-right
in tmux, tmux counts a unicode char as a single char, so using awk
instead of rg/ag fixes this. This also improves portability